tests/requirements-test.txt
# Testing dependencies for AngelaMCP

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0

# Test utilities
pytest-xdist>=3.3.0  # Parallel test execution
pytest-html>=3.2.0   # HTML test reports
pytest-benchmark>=4.0.0  # Performance benchmarking

# Coverage and quality
coverage>=7.2.0
pytest-cov>=4.1.0

# Mocking and fixtures
responses>=0.23.0
factory-boy>=3.3.0
faker>=19.0.0

# Database testing
pytest-postgresql>=5.0.0  # For PostgreSQL testing
aiosqlite>=0.19.0  # For async SQLite testing

# UI testing
pytest-rich>=0.1.0

# Performance and memory testing
memory-profiler>=0.61.0
psutil>=5.9.0

# Test data and fixtures
freezegun>=1.2.0  # Time mocking
testfixtures>=7.1.0

# Development and debugging
pytest-sugar>=0.9.7  # Better test output
pytest-clarity>=1.0.1  # Better assertion output
pytest-picked>=0.4.6  # Run tests based on git changes